Conditionalize the setting of JMX args and remove defunct settings.
Conditionalize the setting of the default JVM arguments for remote JMX connectivity such that setting USE_CUSTOM_JMX_CONNECTION=true would cause none of the JMX args to be set.  Customized arguments can be passed in by adding them to the JAVA_OPTS env var as needed.  This way JMX connectivity can be made to conform to an environment's security and networking requirements.